+++ src/front/pages/Layout.jsx
@@ -4,6 +4,8 @@ import  Footer  from "../components/Footer"
 import NavbarMenu from "../components/Navbar"
 import Body from "../components/Body";
 import Jumbotron from "../components/Jumbotron"
+import ListProduct from "./ListProducts"
+import Products from "./Products"
 
 // Base component that maintains the navbar and footer throughout the page and the scroll to top functionality.
 export const Layout = () => {
@@ -11,10 +13,10 @@ export const Layout = () => {
         <ScrollToTop>
             <NavbarMenu />
             <Jumbotron/>
-            <Body/>
+            <Outlet/>
             <Footer />
-            <ListProduct/>
-            <Products/>
+ {/**           <ListProduct/>
+            <Products/>*/}
         </ScrollToTop>
     )
 }
